Move dialogs from fragments package to dialogs pacakge.
