+++ app/src/main/java/de/danoeh/antennapod/dialog/preferences/PreferenceAutoCompleteTextDialog.java
@@ -1,4 +1,4 @@
-package de.danoeh.antennapod.dialog;
+package de.danoeh.antennapod.dialog.preferences;
 
 import android.app.Dialog;
 import android.os.Bundle;
@@ -32,17 +32,12 @@ import io.reactivex.Observable;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.schedulers.Schedulers;
 
-public class TagDialog extends DialogFragment {
+public class PreferenceAutoCompleteTextDialog extends DialogFragment {
     public static final String TAG = "TagSettingsDialog";
     private static final String ARG_FEED_PREFERENCES = "feed_preferences";
     private List<String> displayedTags;
     private @NonNull TagDialogBinding viewBinding;
-    private TagSelectionAdapter adapter;
-    private List<Feed> feeds;
 
-    public TagDialog(List<Feed> feeds) {
-        this.feeds = feeds;
-    }
 
     @NonNull
     @Override
@@ -52,8 +47,6 @@ public class TagDialog extends DialogFragment {
         displayedTags.remove(FeedPreferences.TAG_ROOT);
 
         viewBinding = TagDialogBinding.inflate(getLayoutInflater());
-        adapter = new TagSelectionAdapter();
-        adapter.setHasStableIds(true);
 
         loadTags();
         viewBinding.newTagEditText.setThreshold(1);
@@ -79,18 +72,24 @@ public class TagDialog extends DialogFragment {
         return dialog.create();
     }
 
+    interface AutoCompleteTextCallback {
+        List<String> loadAutoCompleteText();
+    }
+    AutoCompleteTextCallback autoCompleteTextCallback;
     private void loadTags() {
         Observable.fromCallable(
                 () -> {
-                    NavDrawerData data = DBReader.getNavDrawerData();
-                    List<NavDrawerData.DrawerItem> items = data.items;
-                    List<String> folders = new ArrayList<String>();
-                    for (NavDrawerData.DrawerItem item : items) {
-                        if (item.type == NavDrawerData.DrawerItem.Type.FOLDER) {
-                            folders.add(item.getTitle());
-                        }
-                    }
-                    return folders;
+//                    NavDrawerData data = DBReader.getNavDrawerData();
+//                    List<NavDrawerData.DrawerItem> items = data.items;
+//                    List<String> folders = new ArrayList<String>();
+//                    for (NavDrawerData.DrawerItem item : items) {
+//                        if (item.type == NavDrawerData.DrawerItem.Type.FOLDER) {
+//                            folders.add(item.getTitle());
+//                        }
+//                    }
+//                    return folders;
+
+                    return  autoCompleteTextCallback.loadAutoCompleteText();
                 })
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
@@ -110,46 +109,6 @@ public class TagDialog extends DialogFragment {
         }
         displayedTags.add(name);
         viewBinding.newTagEditText.setText("");
-        adapter.notifyDataSetChanged();
     }
 
-    public class TagSelectionAdapter extends RecyclerView.Adapter<TagSelectionAdapter.ViewHolder> {
-
-        @Override
-        @NonNull
-        public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-            Chip chip = new Chip(getContext());
-            chip.setCloseIconVisible(true);
-            chip.setCloseIconResource(R.drawable.ic_delete);
-            return new ViewHolder(chip);
-        }
-
-        @Override
-        public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
-            holder.chip.setText(displayedTags.get(position));
-            holder.chip.setOnCloseIconClickListener(v -> {
-                displayedTags.remove(position);
-                notifyDataSetChanged();
-            });
-        }
-
-        @Override
-        public int getItemCount() {
-            return displayedTags.size();
-        }
-
-        @Override
-        public long getItemId(int position) {
-            return displayedTags.get(position).hashCode();
-        }
-
-        public class ViewHolder extends RecyclerView.ViewHolder {
-            Chip chip;
-
-            ViewHolder(Chip itemView) {
-                super(itemView);
-                chip = itemView;
-            }
-        }
-    }
 }
