+++ app/src/main/java/de/danoeh/antennapod/dialog/preferences/PreferenceListDialog.java
@@ -1,50 +1,42 @@
-package de.danoeh.antennapod.fragment.preferences.dialog;
+package de.danoeh.antennapod.dialog.preferences;
 
 import android.content.Context;
-import android.view.LayoutInflater;
-import android.view.View;
 
 import androidx.appcompat.app.AlertDialog;
-import androidx.appcompat.widget.SwitchCompat;
 
 import de.danoeh.antennapod.R;
 
-public class PreferenceSwitchDialog {
+public class PreferenceListDialog {
     protected Context context;
     private String title;
-    private String text;
     private OnPreferenceChangedListener onPreferenceChangedListener;
+    private int selectedPos = 0;
 
-    public PreferenceSwitchDialog(Context context, String title, String text) {
+    public PreferenceListDialog(Context context, String title) {
         this.context = context;
         this.title = title;
-        this.text = text;
     }
 
     public interface OnPreferenceChangedListener {
         /**
          * Notified when user confirms preference
          *
-         * @param enabled The preference
+         * @param pos The index of the item that was selected
          */
 
-        void preferenceChanged(boolean enabled);
+        void preferenceChanged(int pos);
     }
 
-    public void openDialog() {
+    public void openDialog(String[] items) {
 
         AlertDialog.Builder builder = new AlertDialog.Builder(context);
         builder.setTitle(title);
-
-        LayoutInflater inflater = LayoutInflater.from(this.context);
-        View layout = inflater.inflate(R.layout.dialog_switch_preference, null, false);
-        SwitchCompat switchButton = layout.findViewById(R.id.dialogSwitch);
-        switchButton.setText(text);
-        builder.setView(layout);
-
+        builder.setSingleChoiceItems(items, selectedPos, (dialog, which) -> {
+            selectedPos = which;
+        });
         builder.setPositiveButton(R.string.confirm_label, (dialog, which) -> {
-            if (onPreferenceChangedListener != null) {
-                onPreferenceChangedListener.preferenceChanged(switchButton.isChecked());
+            if (onPreferenceChangedListener != null && selectedPos >= 0) {
+                onPreferenceChangedListener.preferenceChanged(selectedPos);
             }
         });
         builder.setNegativeButton(R.string.cancel_label, null);
